fix: change import location based on directory refactor
